Fix some bugs for sonar cloud
