+++ src/AltinnCore/Common/Helpers/WorkflowHelper.cs
@@ -70,7 +70,7 @@ namespace AltinnCore.Common.Helpers
                 case WorkflowStep.Archived:
                     return $"/runtime/{applicationOwner}/{applicationId}/{instanceId}/#Preview";
                 default:
-                    return $"/runtime/{applicationOwner}/{applicationId}/ManualTesting";
+                    return $"/designer/{applicationOwner}/{applicationId}/ManualTesting";
             }
         }
