+++ src/AltinnCore/Designer/Controllers/ModelController.cs
@@ -77,25 +77,16 @@ namespace AltinnCore.Designer.Controllers
 
             ServiceMetadata serviceMetadata = null;
 
-            bool useOldParser = false;
-            if (useOldParser)
-            {
-                var seresParser = new SeresXsdParser(_repository);
-                serviceMetadata = seresParser.ParseXsdToServiceMetadata(org, service, mainXsd, null);
-            }
-            else
-            {
-                xsdMemoryStream.Position = 0;
-                reader = XmlReader.Create(xsdMemoryStream, new XmlReaderSettings { IgnoreWhitespace = true });
+            xsdMemoryStream.Position = 0;
+            reader = XmlReader.Create(xsdMemoryStream, new XmlReaderSettings { IgnoreWhitespace = true });
 
-                XsdToJsonSchema xsdToJsonSchemaConverter = new XsdToJsonSchema(reader, _loggerFactory.CreateLogger<XsdToJsonSchema>());
-                JsonSchema schemaJsonSchema = xsdToJsonSchemaConverter.AsJsonSchema();
+            XsdToJsonSchema xsdToJsonSchemaConverter = new XsdToJsonSchema(reader, _loggerFactory.CreateLogger<XsdToJsonSchema>());
+            JsonSchema schemaJsonSchema = xsdToJsonSchemaConverter.AsJsonSchema();
 
-                JsonSchemaToInstanceModelGenerator converter = new JsonSchemaToInstanceModelGenerator(org, service, schemaJsonSchema);
-                serviceMetadata = converter.GetServiceMetadata();
+            JsonSchemaToInstanceModelGenerator converter = new JsonSchemaToInstanceModelGenerator(org, service, schemaJsonSchema);
+            serviceMetadata = converter.GetServiceMetadata();
 
-                HandleTexts(org, service, converter.GetTexts());
-            }
+            HandleTexts(org, service, converter.GetTexts());
 
             if (_repository.CreateModel(org, service, serviceMetadata, mainXsd))
             {
