--- src/AltinnCore/Runtime/Views/ManualTesting/Users.cshtml
@@ -1,19 +1,0 @@
-@model List<AltinnCore.ServiceLibrary.Models.Testdata>
-@using AltinnCore.ServiceLibrary.Models
-@using System.Net
-
-@{ViewBag.Title = "Velg testbruker";}
-
-<h3>Velg testbruker</h3>
-
-<ul class="list-group action-list-group">
-    @foreach (Testdata data in Model)
-    {
-        <li class="list-group-item border-0">
-            <a href="/runtime/ManualTesting/LoginTestUser?id=@data.DataId&ReturnUrl=@WebUtility.UrlEncode(ViewBag.ReturnUrl)" class="list-group-item-action">
-                <div><strong>@data.Title</strong></div>
-                <span class="text-muted">@data.Description</span>
-            </a>
-        </li>
-    }
-</ul>
