+++ src/AltinnCore/Runtime/appsettings.json
@@ -32,8 +32,7 @@
     "RuntimeMode": "AltinnStudio",
     "LanguageFilesLocation": "../Common/Languages/ini/",
     "SoftValidationPrefix": "*WARNING*",
-    "AltinnStudioEndpoint": "http://altinn3.no/",
-    "AuthenticationMode" :  "JwtCookie"
+    "AltinnStudioEndpoint": "http://altinn3.no/"
   },
   "PlatformSettings": {
     "ApiBaseEndpoint": "http://platform.altinn.cloud/api/"
